app/requirements.txt
fastapi
uvicorn
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.1.0
pyyaml>=6.0
pillow
requests
opencv-python-headless
numpy
